Add x64 project configs & 64-bit fixes
Enable x64 builds and improve 64-bit portability across the codebase. Key changes:

- Add Debug/Release x64 ProjectConfiguration entries and make vcxproj configs platform-aware (AudioXA2, DisplayD3D12, GCQ, etc.). Use $(OutDir)/$(IntDir) for outputs and PCH paths to avoid hardcoded Win32 paths and collisions with x64.
- Relax project Condition attributes from combined '$(Configuration)|$(Platform)' checks to separate configuration/platform-aware checks so both Win32 and x64 targets are supported.
- Implement 64-bit-safe type and cast changes: replace dword-based request IDs and hashes with uintptr_t, use uintptr_t for module bases and address arithmetic, and add explicit int casts for strlen/size values to silence warnings and avoid narrowing.
- Improve crash/debug handling for x64: replace inline asm int 3 with __debugbreak; stub out CONTEXT-derived stack-walking on x64 (TODO: port to StackWalk64/RtlCaptureStackBackTrace) and add x64 register formatting in exception reports.
- Misc fixes: cast size in _write call, replace hardcoded import/library paths with $(OutDir), and remove several diagnostic TRACE/log blocks from DisplayD3D12 primitives.
- Add/update third-party assets (stb_image, stb_image_write) and multiple zlib/third-party file updates.

These changes prepare the build system and code for x64 targets and reduce assumptions about 32-bit pointers; x64 stack unwinding is intentionally stubbed and should be replaced with proper StackWalk64/RtlCaptureStackBackTrace later.

+++ Debug/Log.cpp
@@ -238,7 +238,7 @@ int FileReactor::OutputThread::run()
 			m_OutputLock.unlock();
 
 #if defined(_WIN32)
-			_write( fp, sWrite.c_str(), sWrite.size() );
+			_write( fp, sWrite.c_str(), (unsigned int)sWrite.size() );
 #else
 			if ( ::write( fp, sWrite.c_str(), sWrite.size() ) != sWrite.size() )
 				fprintf( stderr, "Failed to write: %s\n", sWrite.c_str() );
